requirements.txt
asgiref==3.2.7
beautifulsoup4==4.6.0
certifi==2020.4.5.1
chardet==3.0.4
cloudinary==1.20.0
Django==3.0.5
django-extensions==2.2.8
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
l18n==2018.5
Pillow==5.4.1
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.8.1
wagtail-generic-chooser==0.1.1
webencodings==0.5.1
Willow==1.3
